requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
requests==2.31.0
stripe==7.8.0
celery==5.3.4
redis==5.0.1
django-ckeditor==6.7.0
django-taggit==4.0.0
django-simple-history==3.4.0
django-cleanup==8.1.0
django-storages==1.14.2
boto3==1.34.0
psycopg2-binary==2.9.9
mysqlclient==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0

